backend/requirements.txt
# Core FastAPI and WebSocket dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
pydantic==2.5.0

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
asyncpg==0.29.0

# AI and Machine Learning libraries
google-genai==0.3.0
openai==1.3.7

# Data processing and utilities
python-dotenv==1.0.0
pydub==0.25.1

# Additional utility libraries
pytz==2023.3
python-dateutil==2.8.2

# HTTP and API utilities
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1

# Audio and media processing
pillow==10.1.0
opencv-python==4.8.1.78

# Redis for caching (optional)
redis==5.0.1
aioredis==2.0.1

# Scheduling and background tasks
apscheduler==3.10.4

# Development and debugging
pytest==7.4.3
black==23.11.0
flake8==6.1.0
